Refactor team-colour lookup to avoid errors
This avoids errors by construction, rather than needing additional
checks. It also has the effect that a corner which doesn't have a
team allocated behaves like a corner with a team which didn't show
up, which seems reasonable (and as far as I can tell follows the
way that this has worked in the past -- this code was modelled on
how the SR2016 compstate scorer does its colour lookups).

This fixes https://github.com/srobo/sr2019-comp/issues/3